II Avalonia 1.9.9.2: Bugfixes - Threading bugfixes re: locking objects  - Including List<PointD> Points and ListPatientEvents - Added additional arbitrary grace period to ClearFuture  - Suspected to be erasing some waveforms as they were being drawn - Transitioned all drawing math from "float" to "double" - Issues fixed w/ bugfixes re: drawing functions  - Fixed Underwrite not working properly d/t logic error   - Resolved #123 in conjunction with ClearFuture grace period - Resolved #124, suspected fixed by ClearFuture grace period - Resolved #125, suspected fixed by ClearFuture grace period

+++ II Core/Classes/Waveform.Dictionary.Functions.cs
@@ -2,20 +2,24 @@
 using System.Collections.Generic;
 
 namespace II.Waveform {
+
     public static partial class Dictionary {
+
         public class Plot {
             public int DrawResolution;
             public int IndexOffset;
-            public float [] Vertices;
+            public double [] Vertices;
 
-            public Plot () { }
+            public Plot () {
+            }
+
             public Plot (int drawResolution, int indexOffset) {
                 DrawResolution = drawResolution;
                 IndexOffset = indexOffset;
             }
         }
 
-        public static Plot Lerp (Plot _Plot1, Plot _Plot2, float _Percent) {
+        public static Plot Lerp (Plot _Plot1, Plot _Plot2, double _Percent) {
             /* Creates a Plot with a lerp of all Y axis points
              * Note: IndexOffset and DrawResolution are only averaged; loss of accuracy possible
              */
@@ -24,7 +28,7 @@ namespace II.Waveform {
                 (_Plot1.DrawResolution + _Plot2.DrawResolution) / 2,
                 (_Plot1.IndexOffset + _Plot2.IndexOffset) / 2);
 
-            List<float> vertices = new List<float> ();
+            List<double> vertices = new List<double> ();
 
             for (int i = 0; i < _Plot1.Vertices.Length || i < _Plot2.Vertices.Length; i++) {
                 if (i < _Plot1.Vertices.Length && i < _Plot2.Vertices.Length)
